scDINO run (#25)
* ready to run

* run preprocessing on HPC

* rerun scDINO

* fix metadata extraction

* addressing comments prior to merge

+++ 6.scDINO_analysis/0.pre-process_images/scripts/0.pre-process_images.py
@@ -11,8 +11,9 @@
 
 
 import multiprocessing as mp
+import os
 import pathlib
-from typing import Optional
+from typing import List, Optional, Tuple
 
 import cv2
 
@@ -23,40 +24,39 @@ import pandas as pd
 import tifffile as tiff
 import tqdm
 
-# ## Import paths
-
-# In[2]:
-
-
-# set the path to the data directory
-data_file_dir = pathlib.Path(
-    "../../../4.process_CP_features/data/annotated_data/run_20231017ChromaLive_6hr_4ch_MaxIP_sc.parquet"
-).resolve(strict=True)
-
-# read in the data
-cp_feature_data = pd.read_parquet(data_file_dir)
-# print the data
-print(cp_feature_data.shape)
-cp_feature_data.head()
+try:
+    cfg = get_ipython().config
+    in_notebook = True
+except NameError:
+    in_notebook = False
 
 
-# In[3]:
+# In[ ]:
 
 
-# get columns that contain Metadata
-metadata_columns = [col for col in cp_feature_data.columns if "Metadata" in col]
-metadata_df = cp_feature_data[metadata_columns]
-# get columns that contain Features
-feature_df = cp_feature_data.drop(columns=metadata_columns)
-metadata_df.head()
+def get_crop_counts(list_of_counts: List[Tuple[int, int, int]]) -> Tuple[int, int, int]:
+    """
+    Get the total counts of successful crops and ommited crops
 
+    Parameters
+    ----------
+    list_of_counts : List[Tuple[int, int, int]]
+        A list of tuples containing the counts of successful crops and ommited crops
 
-# This cell is not run as it takes a long time to run...
-
-# In[4]:
-
-
-# define the function to get the metadata and crop the image
+    Returns
+    -------
+    Tuple[int, int, int]
+        A tuple containing the total counts of successful crops and ommited crops
+    """
+    total_ommited = 0
+    total_sucessful = 0
+    total_total = 0
+    for ommited, sucessful, total in list_of_counts:
+        total_ommited += ommited
+        total_sucessful += sucessful
+        total_total += total
+    assert total_total == total_ommited + total_sucessful
+    return (total_ommited, total_sucessful, total_total)
 
 
 def crop_image(
@@ -65,8 +65,9 @@ def crop_image(
     radius: int = 50,
     add_channels: Optional[bool] = False,
     total_channels: int = 5,
-) -> None:
+) -> Tuple[int, int, int]:
     """
+
     Crop the image based on the metadata and save the cropped image to disk
     Also output extracted metadata for the cropped image
 
@@ -83,27 +84,35 @@ def crop_image(
 
     Returns
     -------
-    None
+    Tuple[int, int, int]
+        A tuple containing the counts of omitted crops, successful crops and total crops
     """
+    successful_count = 0
+    omitted_count = 0
+    total_count = 1
     image_information_df = metadata_df.copy().iloc[i]
 
     center_y = image_information_df["Metadata_Nuclei_Location_Center_Y"].astype(int)
     center_x = image_information_df["Metadata_Nuclei_Location_Center_X"].astype(int)
+    well_fov = image_information_df["Metadata_Well_FOV"]
+    image_path = pathlib.Path(f"{str(image_path)}{well_fov}").resolve(strict=True)
     # DNA
     image_name_DNA = pathlib.Path(image_information_df["Metadata_Image_FileName_DNA"])
     image_path_DNA = pathlib.Path(image_path / image_name_DNA).resolve(strict=True)
     # 488_1
     image_name_488_1 = pathlib.Path(
-        image_information_df["Metadata_Image_FileName_488_1"]
+        image_information_df["Metadata_Image_FileName_CL_488_1"]
     )
     image_path_488_1 = pathlib.Path(image_path / image_name_488_1).resolve(strict=True)
     # 488_2
     image_name_488_2 = pathlib.Path(
-        image_information_df["Metadata_Image_FileName_488_2"]
+        image_information_df["Metadata_Image_FileName_CL_488_2"]
     )
     image_path_488_2 = pathlib.Path(image_path / image_name_488_2).resolve(strict=True)
     # 561
-    image_name_561 = pathlib.Path(image_information_df["Metadata_Image_FileName_561"])
+    image_name_561 = pathlib.Path(
+        image_information_df["Metadata_Image_FileName_CL_561"]
+    )
     image_path_561 = pathlib.Path(image_path / image_name_561).resolve(strict=True)
     image_DNA = tiff.imread(image_path_DNA)
     image_488_1 = tiff.imread(image_path_488_1)
@@ -127,11 +136,8 @@ def crop_image(
     # Where edge case is cells that are too close to the edge of the image to crop
     # This ensures that all crops are the same dimensions and can be used in the model
     if image_DNA_crop.shape[0] < radius * 2 or image_DNA_crop.shape[1] < radius * 2:
-        print(
-            f"Image {image_information_df['Metadata_Image_FileName_DNA']} is an edge case. Ommitting..."
-        )
-        ommited_couunt = 1
-        return None
+        ommited_count = 1
+        return (omitted_count, successful_count, total_count)
     # merge the channels to a single image
     image_merge = np.stack(
         [image_DNA_crop, image_488_1_crop, image_488_2_crop, image_561_crop], axis=-1
@@ -145,49 +151,94 @@ def crop_image(
                 image_merge = np.concatenate((image_merge, new_channels), axis=-1)
     # save images to disk
     image_save_path = pathlib.Path(
-        f"../../../data/processed_images/crops/{image_information_df['Metadata_Well']}/image_number_{image_information_df['Metadata_ImageNumber']}_cell_number_{image_information_df['Metadata_Nuclei_Number_Object_Number']}/"
+        # f"../../../data/processed_images/crops/{image_information_df['Metadata_Well_FOV']}/time_{image_information_df['Metadata_Time']}_image_number_{image_information_df['Metadata_ImageNumber']}_cell_number_{image_information_df['Metadata_Nuclei_Number_Object_Number']}/"
+        f"../data/processed_images/sc_crops/{image_information_df['Metadata_Well_FOV']}_time_{image_information_df['Metadata_Time']}_image_number_{image_information_df['Metadata_ImageNumber']}_cell_number_{image_information_df['Metadata_Nuclei_Number_Object_Number']}/"
     )
     image_save_path.mkdir(parents=True, exist_ok=True)
     file_name = image_information_df["Metadata_Image_FileName_DNA"].replace(
         ".tiff",
-        f'image_number_{image_information_df["Metadata_ImageNumber"]}_{image_information_df["Metadata_Nuclei_Number_Object_Number"]}_crop.tiff',
+        f'image_number_{image_information_df["Metadata_ImageNumber"]}_cell_number_{image_information_df["Metadata_Nuclei_Number_Object_Number"]}_crop.tiff',
     )
     image_save_path = pathlib.Path(image_save_path / file_name)
+    if os.path.exists(image_save_path):
+        sucessful_count = 1
+        return (omitted_count, successful_count, total_count)
     tiff.imwrite(image_save_path, image_merge)
-    return None
+    successful_count = 1
+    return (omitted_count, successful_count, total_count)
 
 
-# In[5]:
+# ## Import paths
+
+# In[3]:
+
+
+# set the path to the data directory
+data_file_dir = pathlib.Path(
+    "../../../5.process_CP_features/data/4.normalized_data/profiles/normalized_profile.parquet"
+).resolve(strict=True)
+
+# read in the data
+cp_feature_data = pd.read_parquet(data_file_dir)
+# print the data
+print(cp_feature_data.shape)
+cp_feature_data.head()
+
+
+# In[4]:
+
+
+well_fov = cp_feature_data["Metadata_Well"] + "_F" + cp_feature_data["Metadata_FOV"]
+cp_feature_data.insert(0, "Metadata_Well_FOV", well_fov)
+# get columns that contain Metadata
+metadata_columns = [col for col in cp_feature_data.columns if "Metadata" in col]
+metadata_df = cp_feature_data[metadata_columns]
+# get columns that contain Features
+feature_df = cp_feature_data.drop(columns=metadata_columns)
+# show all columns
+metadata_df.head()
+
+
+# This cell is not run as it takes a long time to run...
+
+# In[ ]:
 
 
 image_path = pathlib.Path(
-    "../../../2.cellprofiler_ic_processing/illum_directory/20231017ChromaLive_6hr_4ch_MaxIP/"
+    "../../../2.cellprofiler_ic_processing/illum_directory/timelapse/20231017ChromaLive_6hr_4ch_MaxIP_"
 )
 radius = 50
 
-
-# ### notebook is stopped here to run the following cell as a script
 
 # In[ ]:
 
 
 # set the number of processes to use
-num_processes = mp.cpu_count() - 2
+if in_notebook:
+    num_processes = mp.cpu_count() - 4
+else:
+    num_processes = mp.cpu_count()
 print(f"Number of processes: {num_processes}")
 # get the total number of rows in the metadata_df
 total_crops = metadata_df.shape[0]
-print(f"There are {total_crops} images to crop")
+print(f"There are {total_crops:,} images to crop")
 
 process_list = [
     mp.Process(target=crop_image, args=(i, image_path, radius, False))
     for i in range(total_crops)
 ]
-print(f"There are {len(process_list)} processes to run")
+print(f"There are {len(process_list):,} processes to run")
 pool = mp.Pool(num_processes)
 # run the processes in the pool with multiple args
-pool.starmap_async(
+results = pool.starmap_async(
     crop_image, [(i, image_path, radius, False) for i in range(total_crops)]
 )
 pool.close()
 pool.join()
 pool.terminate()
+results = results.get()
+final_counts = get_crop_counts(results)
+# print the totals with commas for easy reading
+print(
+    f"Total crops: {final_counts[2]:,}, Sucessful crops: {final_counts[1]:,}, Ommited crops: {final_counts[0]:,}"
+)
